CB first cut at abstracting components
