+++ package.json
@@ -11,7 +11,9 @@
     "react": "^15.4.2",
     "react-addons-update": "^15.4.2",
     "react-bootstrap": "^0.30.7",
-    "react-dom": "^15.4.2"
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.3",
+    "redux": "^3.6.0"
   },
   "scripts": {
     "start": "react-scripts start",
