+++ src/helpers.js
@@ -1,0 +1,30 @@
+// shuffle array
+export const shuffleArray = function(array) {
+    for (var i = array.length - 1; i > 0; i--) {
+        var j = Math.floor(Math.random() * (i + 1));
+        var temp = array[i];
+        array[i] = array[j];
+        array[j] = temp;
+    }
+    return array;
+}
+
+//Finds y value of given object
+export const findPos = function(obj) {
+    var curtop = 0;
+    if (obj.offsetParent) {
+        do {
+            curtop += obj.offsetTop;
+        } while (obj = obj.offsetParent);
+    return [curtop];
+    }
+}
+
+export const arrOfLetters = 'a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p,q,r,s,t,u,v,w,x,y,z'.split(',');
+export const wrongText = [
+  'Sorry, that\'s incorrect.',
+  'Good try, but that\'s incorrect.',
+  'Oops. Good try, though.',
+  'That\'s not it.',
+  'Nope. It\'s a different answer.'
+];
